SpoonEmail: encapsulate recipients with quotes.
If a recipient header contains a comma it will be treated as seperate
recipients. As described in http://tools.ietf.org/html/rfc2822#section-3.6.3 .

Example:
"To: Spoon, Library <spoon@spo.on>" will be treated as 2 recipients "Spoon" and "Library".

To prevent this we should encapsulate the name of a recipient in quotes.