Simulate behaviour of PHP < 5.4
In PHP 5.4, casting an array to a string throws a notice;

    "PHP Notice:  Array to string conversion in ... "

Because Spoon uses casting all over the place, be it for defensive
programming or other reason, these casts go wrong in PHP 5.4.

In this change we touch as little as possible to Spoon's code base,
trying to preserve its whacked behaviour under PHP 5.4.