Bugfix: the token-check would always resolve
When the $useToken-parameter is true, a field is added by the code. In the code
that validates the token the field was used, but the return of getValue() would
return the value that was set during creating the field instead of the one in
POST/GET.

Therefore we use from now on the raw POST/GET-data to check if a token is
available.