Mark a form as incorrect when no CSRF token is present in either the GET or POST parameters.
This fixes a security issue where Spoon would only check if a token was present in the session,
and a form_token field was present in the form. This would always validate as correct, since
the form_token field is added in the constructor of the form with the value stored in the session
if it exists. This allowed CSRF attacks to go through. The missing piece of the CSRF protection
is to also validate against the $_GET and $_POST parameters. If no form_token parameter is present
in those, then we're dealing with a CSRF attack.